reorganize subsystem
