+++ src/main/java/frc/robot/RobotContainer.java
@@ -54,15 +54,15 @@ import frc.robot.commands.util.MatchtimeStatusCommand;
 import frc.robot.commands.util.ShiftAlertingCommand;
 import frc.robot.commands.util.VoltageStatusCommand;
 import frc.robot.Constants.DriveConstants;
-import frc.robot.subsystems.IndexerSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.subsystems.IntakeSubsystem.IntakePreset;
 import frc.robot.subsystems.Limelight;
 import frc.robot.subsystems.Limelight.LimelightType;
-import frc.robot.subsystems.LoaderSubsystem;
-import frc.robot.subsystems.SwerveSubsystem;
-import frc.robot.subsystems.TurretSubsystem;
-import frc.robot.subsystems.TurretSubsystem.TurretTargets;
+import frc.robot.subsystems.drive.SwerveSubsystem;
+import frc.robot.subsystems.manipulator.IndexerSubsystem;
+import frc.robot.subsystems.manipulator.IntakeSubsystem;
+import frc.robot.subsystems.manipulator.IntakeSubsystem.IntakePreset;
+import frc.robot.subsystems.manipulator.LoaderSubsystem;
+import frc.robot.subsystems.manipulator.TurretSubsystem;
+import frc.robot.subsystems.manipulator.TurretSubsystem.TurretTargets;
 import frc.robot.util.AllianceFlipUtil;
 import frc.robot.util.LimelightContainer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
