+++ src/main/java/frc/robot/commands/control/ManualTurretCommand.java
@@ -4,16 +4,15 @@ import static edu.wpi.first.units.Units.*;
 
 import frc.robot.Constants.MetaConstants;
 import frc.robot.Constants.TurretConstants;
-import frc.robot.subsystems.TurretSubsystem;
+import frc.robot.subsystems.manipulator.TurretSubsystem;
+import frc.robot.subsystems.drive.SwerveSubsystem;
 
 
 import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularVelocity;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.SwerveSubsystem;
 
 public class ManualTurretCommand extends Command {
     private final TurretSubsystem m_LauncherSubsystem;
