+++ src/main/java/frc/robot/subsystems/manipulator/IndexerSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.manipulator;
 
 import static edu.wpi.first.units.Units.*;
